@@ -1,29 +1,38 @@
-source 'http://rubygems.org'
+source 'https://rubygems.org'
 
-gem 'rails', '~> 3.1'
-gem 'cancan'
-gem 'devise', '~> 1.4'
-gem 'rails_admin', :git => 'git://github.com/sferik/rails_admin.git'
-gem 'mlb', '~> 0.5'
-gem 'paperclip', '~> 2.4'
-gem 'mini_magick'
-gem 'carrierwave'
-gem 'dragonfly'
-gem 'haml'
-gem 'taps'
-gem "ckeditor", "3.7.2"
-gem 'heroku'
-group :production do
-  gem 'pg'
-end
-group :development, :test do
-  gem 'nifty-generators'
-  gem 'sqlite3'
-end
+gem 'rails', '3.2.9'
+
+# Bundle edge Rails instead:
+# gem 'rails', :git => 'git://github.com/rails/rails.git'
+
+gem 'sqlite3'
+
+
+# Gems used only for assets and not required
+# in production environments by default.
 group :assets do
-  gem 'sass-rails', '~> 3.1'
-  gem 'coffee-rails', '~> 3.1'
-  gem 'uglifier', '~> 1.0'
+  gem 'sass-rails',   '~> 3.2.3'
+  gem 'coffee-rails', '~> 3.2.1'
+
+  # See https://github.com/sstephenson/execjs#readme for more supported runtimes
+  # gem 'therubyracer', :platforms => :ruby
+
+  gem 'uglifier', '>= 1.0.3'
 end
 
-gem "mocha", :group => :test
+gem 'jquery-rails'
+
+# To use ActiveModel has_secure_password
+# gem 'bcrypt-ruby', '~> 3.0.0'
+
+# To use Jbuilder templates for JSON
+# gem 'jbuilder'
+
+# Use unicorn as the app server
+# gem 'unicorn'
+
+# Deploy with Capistrano
+# gem 'capistrano'
+
+# To use debugger
+# gem 'debugger'
