@@ -1,8 +1,8 @@
 # Be sure to restart your server when you modify this file.
 
-DummyApp::Application.config.session_store :cookie_store, :key => '_dummy_app_session'
+Processo::Application.config.session_store :cookie_store, key: '_processo_session'
 
 # Use the database for sessions instead of the cookie-based default,
 # which shouldn't be used to store highly confidential information
 # (create the session table with "rails generate session_migration")
-# DummyApp::Application.config.session_store :active_record_store
+# Processo::Application.config.session_store :active_record_store
