@@ -4,4 +4,4 @@
 # If you change this key, all old signed cookies will become invalid!
 # Make sure the secret is at least 30 characters and all random,
 # no regular words or you'll be exposed to dictionary attacks.
-DummyApp::Application.config.secret_token = '2e733773e594846c6bb87fe0d0bb8e70ea81d334fccdbceca5b05ede2e61f10ea98add53783bf333d8d5523e001f4d7045d71c57ab288e279626c49b18810f9e'
+Processo::Application.config.secret_token = '247be08f42384b6a564f797d6387f26d6af8eea1abc630e2bef0ba164aff4d4c74f65275c7fcac8c430c395e02bb55d7a0598280f403228f76d8b4039392b1c1'
